# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "chess api for playing chess online.",
        "title": "Chess API",
        "contact": {},
        "license": {
            "name": "MIT"
        }
    },
    "paths": {
...
        "/users": {
            "post": {
                "description": "Username can be between 3-20 characters.\nPassword must be at least 3 characters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create an account using provided username and password.",
                "parameters": [
                    {
                        "description": "Register Account",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.UserCredentials"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Api Key",
                        "schema": {
                            "$ref": "#/definitions/server.ApiKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorReason"
                        }
                    },
                    "409": {
                        "description": "Username already exists",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorReason"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorReason"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Must contain ApiKey in the format Bearer: apiKey",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorReason"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorReason"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "server.ApiKeyResponse": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "type": "string"
                }
            }
        },
...
        "server.ErrorReason": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string",
                    "example": "reason"
                }
            }
        },
...
        "server.UserCredentials": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Password123"
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 4,
                    "example": "JohnDoe"
                }
            }
        }
    }
}